Move tests module to separate file
